backend/pyproject.toml
[tool.poetry]
name = "mojbrowarek"
version = "0.0.1"
description = ""
authors = ["Damian Kobyliński <damiankob12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.15.1"
drf-spectacular = "^0.27.2"
typing-extensions = "^4.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
